- Fixed: Terrain shader, thanks to Nadro from the Irrlicht forum.
git-svn-id: http://svn.code.sf.net/p/irrrpgbuilder/code@339 528fbfa9-60f6-4f4a-a82b-d133dc65ca33

+++ src/fx/ShaderCallBack.cpp
@@ -28,35 +28,35 @@ ShaderCallBack* ShaderCallBack::getInstance()
 
 void ShaderCallBack::setFlagEditingTerrain(bool edit)
 {
-    editingTerrain = edit;
+    editingTerrain = (int)edit;
 }
 
 void ShaderCallBack::OnSetConstants(video::IMaterialRendererServices* services, s32 userData)
 {
     layer=0;
-    services->setPixelShaderConstant("terrainLayer0",(float*)&layer,1);
+    services->setPixelShaderConstant("terrainLayer0",(int*)&layer,1);
     layer=1;
-    services->setPixelShaderConstant("terrainLayer1",(float*)&layer,1);
+    services->setPixelShaderConstant("terrainLayer1",(int*)&layer,1);
     layer=2;
-    services->setPixelShaderConstant("terrainLayer2",(float*)&layer,1);
+    services->setPixelShaderConstant("terrainLayer2",(int*)&layer,1);
     layer=3;
-    services->setPixelShaderConstant("terrainLayer3",(float*)&layer,1);
+    services->setPixelShaderConstant("terrainLayer3",(int*)&layer,1);
     layer=4;
-    services->setPixelShaderConstant("terrainLayer4",(float*)&layer,1);
+    services->setPixelShaderConstant("terrainLayer4",(int*)&layer,1);
 
-    services->setPixelShaderConstant("editingTerrain",(float*)&editingTerrain,1);
+    services->setPixelShaderConstant("editingTerrain",(int*)&editingTerrain,1);
 #ifdef EDITOR
 	float plateau = (float)GUIManager::getInstance()->getScrollBarValue(SC_ID_TERRAIN_BRUSH_PLATEAU);
 	services->setPixelShaderConstant("plateau",(float*)&plateau,1);
 #endif
 
     layer=10;
-    services->setPixelShaderConstant("terrainTextureScale",(float*)&layer,1);
+    services->setPixelShaderConstant("terrainTextureScale",(int*)&layer,1);
 	// Retrieve the scale of the terrain
 	
 	layer=(int)TerrainManager::getInstance()->getTileMeshSize();
 	
-	services->setPixelShaderConstant("terrainScale",(float*)&layer,1);
+	services->setPixelShaderConstant("terrainScale",(int*)&layer,1);
 		
 	//services->setPixelShaderConstant("terrainScale",(float*)&layer,1);
 
@@ -64,9 +64,9 @@ void ShaderCallBack::OnSetConstants(video::IMaterialRendererServices* services, 
     //services->setPixelShaderConstant("fogDistance",(float*)&layer,1);
 
     layer=0;
-    services->setPixelShaderConstant("oceanNormalMap",(float*)&layer,1);
+    services->setPixelShaderConstant("oceanNormalMap",(int*)&layer,1);
     layer=1;
-    services->setPixelShaderConstant("oceanReflection",(float*)&layer,1);
+    services->setPixelShaderConstant("oceanReflection",(int*)&layer,1);
 
     SColorf color = App::getInstance()->getDevice()->getSceneManager()->getAmbientLight();
     services->setVertexShaderConstant("AmbientLight",reinterpret_cast<f32*>(&color), 4);
